Add typed new declarations and C type mapping

+++ tests/fixtures/codegen_function.c
@@ -1,11 +1,14 @@
 #include <stdio.h>
 
+#include <stdint.h>
+#include <stdbool.h>
+
 int add(int a, int b) {
-    int c = (a + b);
+    int64_t c = (a + b);
     return 0;
 }
 
 int main(void) {
-    int x = 3;
+    int64_t x = 3;
     return 0;
 }
